requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.0
PyYAML==3.11
Werkzeug==0.9.6
cssselect==0.9.1
gunicorn==19.0.0
itsdangerous==0.24
lxml==3.3.5
# -e git+git://github.com/karls/newspaper.git@extraction-improvements#egg=newspaper
newspaper3k==0.1.4
nltk==3.0.1
requests==2.3.0
simplejson==3.5.3
